Split format.ts into string.ts and date.ts
