+++ src/app/auth/_components/SignUpForm.tsx
@@ -135,7 +135,7 @@ export const SignupForm = () => {
                       style={{ top: "50%", transform: "translateY(-50%)" }}
                       className="absolute right-2 text-gray-800 z-10"
                     >
-                      {showPassword ? <EyeOff size={20} /> : <Eye size={20} />}
+                      {showPassword ? <Eye size={20} /> : <EyeOff size={20} />}
                     </button>
                   </div>
 
@@ -168,7 +168,7 @@ export const SignupForm = () => {
                       style={{ top: "50%", transform: "translateY(-50%)" }}
                       className="absolute right-2 text-gray-800 z-10"
                     >
-                      {showPassword ? <EyeOff size={20} /> : <Eye size={20} />}
+                      {showPassword ? <Eye size={20} /> : <EyeOff size={20} />}
                     </button>
                   </div>
                   {fieldState.invalid && (
