+++ src/app/dashboard/regular-foods/_components/_schema/RegularFoodFormSchema.ts
@@ -1,4 +1,4 @@
-import { toHalfWidth } from "@/utils/format";
+import { toHalfWidth } from "@/utils/format/string";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
