+++ src/app/dashboard/target-Kcal-plans/_component/TargetKcalListItem.tsx
@@ -3,7 +3,7 @@
 import { TargetKcalActionMenu } from "@/app/dashboard/target-kcal-plans/_component/TargetKcalActionMenu";
 import { List } from "@/components";
 import { SelectTargetKcalPlansRecord } from "@/db/schema";
-import { formatDateWithDay } from "@/utils/format";
+import { formatDateWithDay } from "@/utils/format/date";
 import { memo } from "react";
 
 type TargetKcalListItemProps = {
