+++ src/app/dashboard/target-Kcal-plans/_component/schema/TargetKcalPlanFormSchema.tsx
@@ -1,4 +1,5 @@
-import { createJstDate, toHalfWidth } from "@/utils/format";
+import { createJstDate } from "@/utils/format/date";
+import { toHalfWidth } from "@/utils/format/string";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
