+++ src/app/setup/_components/_schema/targetKcalFormSchema.ts
@@ -1,4 +1,4 @@
-import { toHalfWidth } from "@/utils/format";
+import { toHalfWidth } from "@/utils/format/string";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
